PHP 7 compatability change.

+++ src/libraries/sessions-cache.php
@@ -16,30 +16,30 @@ class Session_Driver_cache extends Session_Driver
 	protected $_changed = false;
 	protected $_renewable = true;
 	protected $_regenerated = false;
-	
+
 	public function id()
 	{
 		if(!is_null($this->_id))
 			return $this->_id;
-		
+
 		// Let's attempt to find our session identifier using cookies.
 		if(Cookies::has($this->_cookie)) {
 			$this->_id = Cookies::get($this->_cookie)->value;
 			return $this->_id;
 		}
-		
+
 		// If we couldn't find it, let's generate a new one and set the cookie (encrypted, secure, httponly).
 		$this->regenerateID(false);
-		
+
 		// Return the identifier.
 		return $this->_id;
 	}
-	
+
 	protected function regenerateID()
 	{
 		// Set the ID to a random string.
 		$this->_id = $this->_generateID();
-		
+
 		// Send the cookie.
 		$cookie = new Cookie($this->_cookie);
 		$cookie->value = $this->_id;
@@ -50,25 +50,25 @@ class Session_Driver_cache extends Session_Driver
 		$cookie->httponly = true;
 		$cookie->send();
 	}
-	
+
 	protected $_idchars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_';
 	protected function _generateID()
 	/* Returns a unique, randomly-generated session identifier */
 	{
 		// Seed the random number generator.
-		mt_srand(microtime(true) * 1000000);
-		
+		mt_srand((int) (microtime(true) * 1000000));
+
 		// Create a unique session identifier.
 		$length = 100 + mt_rand(25,50);
 		$id = "";
-		
+
 		while(strlen($id) < $length) {
 			$id .= substr($this->_idchars, mt_rand(0, strlen($this->_idchars)-1), 1);
 		}
-		
+
 		return $id;
 	}
-	
+
 	public function defaultVars()
 	{
 		return array(
@@ -78,7 +78,7 @@ class Session_Driver_cache extends Session_Driver
 			'last_regenerate' => 0 // Force regeneration on creation to prevent fixation.
 		);
 	}
-	
+
 	public function load()
 	{
 		// Load our session variables from the cache.
@@ -86,7 +86,7 @@ class Session_Driver_cache extends Session_Driver
 		$data = Cache::section('session')->get( sha1($this->id()), function() use (&$object) {
 			return array( 'vars' => $object->defaultVars(), 'flash' => array() );
 		});
-		
+
 		// Copy to instance if the session is not expired.
 		if(!(isset($data['vars']['expire']) && $data['vars']['expire'] > time())) {
 			$this->_vars = $data['vars'];
@@ -95,11 +95,11 @@ class Session_Driver_cache extends Session_Driver
 		else {
 			$this->_vars = $this->defaultVars();
 		}
-		
+
 		// Update the renewable flag (if applicable).
 		if(isset($data['vars']['renewable']))
 			$this->_renewable = $data['vars']['renewable'];
-		
+
 		// Handle Variable Updates and Session Regeneration
 		$regenerate = false;
 		foreach($this->defaultVars() as $k => $v) {
@@ -110,34 +110,34 @@ class Session_Driver_cache extends Session_Driver
 				$this->set($k, $v);
 			}
 		}
-		
+
 		if((time() - $this->get('last_regenerate')) > 600) {
 			$regenerate = true;
 		}
-		
+
 		if($regenerate)
 			$this->regenerate();
 	}
-	
+
 	public function unload()
 	{
 		// We can save data here (if anything was changed).
-		if($this->_changed || sizeof($this->_flashwrite) > 0 || sizeof($this->_flash) > 0) { 
+		if($this->_changed || sizeof($this->_flashwrite) > 0 || sizeof($this->_flash) > 0) {
 			$this->save();
 		}
 	}
-	
+
 	public function save()
 	{
 		if(!$this->_renewable)
 			return;
-			
+
 		Cache::section('session')->put( sha1($this->id()), array(
 			'vars' => $this->_vars,
 			'flash' => $this->_flashwrite
 		) );
 	}
-	
+
 	public function get($key, $default=null)
 	{
 		if($key == 'id')
@@ -146,34 +146,34 @@ class Session_Driver_cache extends Session_Driver
 			return $this->auth();
 		if($key == 'user')
 			return $this->user();
-			
+
 		if(isset($this->_vars[$key]))
 			return $this->_vars[$key];
-			
+
 		if(isset($this->_flash[$key]))
 			return $this->_flash[$key];
-			
+
 		return value($default);
 	}
-	
+
 	public function set($key, $value)
 	{
 		$this->_changed = true;
 		$this->_vars[$key] = value($value);
 	}
-	
+
 	public function has($key)
 	{
 		return (isset($this->_vars[$key]) || isset($this->_flash[$key]));
 	}
-	
+
 	public function forget($key)
 	{
 		$this->_changed = true;
 		unset($this->_vars[$key]);
 		unset($this->_flash[$key]);
 	}
-	
+
 	public function flash($key, $value=null)
 	{
 		if($value === null && !is_array($key) && isset($this->_vars[$key])) {
@@ -181,14 +181,14 @@ class Session_Driver_cache extends Session_Driver
 			$this->_flashwrite[$key] = $this->_flash[$key];
 			unset($this->_vars[$key]);
 		}
-		
+
 		if(!is_array($key) && $value !== null)
 		{
 			$this->_flash[$key] = value($value);
 			$this->_flashwrite[$key] = $this->_flash[$key];
 			unset($this->_vars[$key]);
 		}
-		
+
 		if(is_array($key))
 		{
 			foreach($key as $k => $v) {
@@ -197,20 +197,20 @@ class Session_Driver_cache extends Session_Driver
 				unset($this->_vars[$k]);
 			}
 		}
-		
+
 		$this->_changed = true;
 	}
-	
+
 	public function keep($key)
 	{
 		$this->_changed = true;
-		
+
 		if(isset($this->_flash[$key])) {
 			$this->_vars[$key] = $this->_flash[$key];
 			unset($this->_flash[$key]);
 		}
 	}
-	
+
 	public function regenerate($saveExisting=true)
 	{
 		if($this->_regenerated === true)
@@ -219,21 +219,21 @@ class Session_Driver_cache extends Session_Driver
 		// Mark the session as invalid (in 60 seconds).
 		$this->_vars['renewable'] = false;
 		$this->_vars['expire'] = time() + 60;
-		
+
 		// We need to make sure to save.
 		if($saveExisting === true)
 			$this->save();
-		
+
 		// Let's mark the session as valid again.
 		unset($this->_vars['expire']);
 		unset($this->_vars['renewable']);
-		
+
 		// Let's change the session identifier and update the cookie.
 		$this->regenerateID();
-		
+
 		// Let's update the variable.
 		$this->_vars['last_regenerate'] = time();
-		
+
 		// Let's save changes.
 		$this->_regenerated = true;
 		$this->save();
